Finish styling for color and font page
I finished all of the stylings that relates to the information that was given on the colors and fonts page of the project. I also am adding this directly to the GitHub site from my computer. I will then add the SSH key to my git bash so that I can upload it from the command line instead of manually on this website.